Delete an equation when it is at the end of the previous line.
Case 1: We are in the line the equation is on and should be able to delete the equation.
Case 2: There is a equation at the end of the previous line, we hit backspace at the start of the line and the equation is delete and the current line is deleted.

This will make Case 2 work.