Refactor Line Merging of multiple lines.
When there is an equation at the end of a line and we are backspacing
from the line below. Instead of trying to treat the equation at the end
of a line as a special case, we place the cursor at the end of the
equation and then let it be handled as the case of deleting a equation
within a line.