Remove text nodes within Mathjax equations.
 We don’t know which node the browser will give us, therefore when
checking if it was part of a Mathjax equation (was an SVG element or
script tag), it’s not a bad idea to check that it’s parent wasn’t
either. For example, the browser may have passed us the text node
within the script. When we’ve found that an element was in an SVG, we
iterate up the DOM tree until we find it’s parent.